Unity Multiplayer Lesson Complete
These past hours were me having major problems since the Lilypads wouldn't spawn on the client side.

Another one was the score not appearing on the client side as well.

I realize now that some Network Identities don't need to be server-only or local player authority. You can leave both unchecked [which is what I did to the Lilypad prefab and the Inventory GameObject in order for the score/pads to synchronize]

With that, I have finally completed this Lesson. It was certainly a lot to take in, but at least with the immense amounts of problems - I think I have a better idea of how the Client - Server - Host relationship works as well as the coding/Editor settings behind this feature.